+++ docs/_static/versions.json
@@ -2,7 +2,7 @@
     "versions": [
         {
             "name": "STABLE",
-            "url": "/fatal: not a git repository (or any of the parent directories): .git/index.html"
+            "url": "/0.6.10/index.html"
         }
     ]
 }
