+++ src/spectrochempy/processing/baselineprocessing/baseline_deprecated.py
@@ -1,7 +1,7 @@
 # DEPRECATED
 
 from spectrochempy.application import warning_
-from spectrochempy.processing.baselineprocessing.baseline import Baseline
+from spectrochempy.processing.baselineprocessing.baselineprocessing import Baseline
 from spectrochempy.utils.decorators import deprecated
 
 __all__ = ["BaselineCorrection", "ab", "abc"]
